Fixed infinite API call bug
Fixed a bug where componentDidUpdate did not have a proper exit
condition, causing it to call the API infinitely. This has been fixed
by adding a boolean 'updated' value to the state that is set to 'true'
after creating, patching, or deleting a resource, and is then set back
to false after running componentDidUpdate.