Initial commit for adding a damage graph
This is the first step in implementing a graph of damage results from
repeated combat phase. The eventual goal is display a graph of the
results of 100 combat phases to show a range of expected results,
rather than a single average damage (as is used in the app now).
Currently, the first step is to get working functions to create data
points in the format that would be accepted by the Recharts graph
library. This required moving over relevant combat functions diceRoll.js